Move handlers from newflow
oauth_callback
handlers/verify_user_email-by_pin
verify_email_by_code
oauth_callback
send_reset_password_email
find_user_by_token
