--- app/handlers/newflow/find_user_by_token.rb
@@ -1,33 +1,0 @@
-module Newflow
-  class FindUserByToken
-    lev_handler
-
-    protected ###############
-
-    def authorized?
-      true
-    end
-
-    def handle
-      if already_logged_in?
-        outputs.user = caller
-        return
-      end
-
-      fatal_error(code: :token_blank) unless params[:token].present?
-
-      user = User.find_by(login_token: params[:token])
-
-      fatal_error(code: :unknown_login_token) unless user.present?
-      fatal_error(code: :expired_login_token) if user.login_token_expired?
-
-      outputs.user = user
-    end
-
-    private #################
-
-    def already_logged_in?
-      !caller.is_anonymous?
-    end
-  end
-end
